Removed use of else conditions.

+++ Command/DeleteCommand.php
@@ -34,27 +34,30 @@ class DeleteCommand extends CacheCommand
         $cacheName     = $input->getArgument('cache-name');
         $cacheProvider = $this->getCacheProvider($cacheName);
         $cacheId       = $input->getArgument('cache-id');
+        $all           = $input->getOption('all');
 
-        if ($input->getOption('all')) {
-            if (method_exists($cacheProvider, 'deleteAll')) {
-                $success = $cacheProvider->deleteAll();
-                $color   = $success ? 'info' : 'error';
-                $success = $success ? 'succeeded' : 'failed';
-                $message = "Deletion of <$color>all</$color> entries in <$color>%s</$color> has <$color>%s</$color>";
-                $output->writeln(sprintf($message, $cacheName, $success, true));
-            } else {
-                throw new \RuntimeException('Cache provider does not implement a deleteAll method.');
-            }
-        } else {
-            if ($cacheId) {
-                $success = $cacheProvider->delete($cacheId);
-                $color   = $success ? 'info' : 'error';
-                $success = $success ? 'succeeded' : 'failed';
-                $message = "Deletion of <$color>%s</$color> in <$color>%s</$color> has <$color>%s</$color>";
-                $output->writeln(sprintf($message, $cacheId, $cacheName, $success, true));
-            } else {
-                throw new \InvalidArgumentException('Missing cache ID');
-            }
+        if ($all && ! method_exists($cacheProvider, 'deleteAll')) {
+            throw new \RuntimeException('Cache provider does not implement a deleteAll method.');
         }
+
+        if ( ! $all && ! $cacheId) {
+            throw new \InvalidArgumentException('Missing cache ID');
+        }
+
+        $success = $all ? $cacheProvider->deleteAll() : $cacheProvider->delete($cacheId);
+        $color   = $success ? 'info' : 'error';
+        $success = $success ? 'succeeded' : 'failed';
+
+        if ( ! $all) {
+            $message = "Deletion of <$color>%s</$color> in <$color>%s</$color> has <$color>%s</$color>";
+            $message = sprintf($message, $cacheId, $cacheName, $success, true);
+        }
+
+        if ($all) {
+            $message = "Deletion of <$color>all</$color> entries in <$color>%s</$color> has <$color>%s</$color>";
+            $message = sprintf($message, $cacheName, $success, true);
+        }
+
+        $output->writeln($message);
     }
 }
